feat(frontend): add notifications for disabled js
